talkieai-server/requirements.txt
uvicorn==0.20.0
azure_storage==0.36.0
fastapi==0.109.0
openai==1.9.0
pydantic==1.10.5
pydub==0.25.1
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.20
starlette==0.35.0
zhipuai==1.0.7
pymysql~=1.0.2
azure-cognitiveservices-speech~=1.34.1
requests~=2.28.2
python-multipart
